+++ common/src/main/java/dev/ftb/mods/ftbxmodcompat/ftbquests/kubejs/KubeJSIntegration.java
@@ -1,6 +1,7 @@
 package dev.ftb.mods.ftbxmodcompat.ftbquests.kubejs;
 
 import dev.architectury.event.EventResult;
+import dev.ftb.mods.ftbquests.api.event.CustomFilterDisplayItemsEvent;
 import dev.ftb.mods.ftbquests.events.CustomRewardEvent;
 import dev.ftb.mods.ftbquests.events.CustomTaskEvent;
 import dev.ftb.mods.ftbquests.events.ObjectCompletedEvent;
@@ -16,11 +17,13 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 
 public class KubeJSIntegration extends KubeJSPlugin {
+	@Override
 	public void init() {
 		CustomTaskEvent.EVENT.register(KubeJSIntegration::onCustomTask);
 		CustomRewardEvent.EVENT.register(KubeJSIntegration::onCustomReward);
 		ObjectCompletedEvent.GENERIC.register(KubeJSIntegration::onCompleted);
 		ObjectStartedEvent.GENERIC.register(KubeJSIntegration::onStarted);
+		CustomFilterDisplayItemsEvent.ADD_ITEMSTACK.register(KubeJSIntegration::onCustomFilterItem);
 
 		Stages.added(event -> {
 			if (event.getPlayer() instanceof ServerPlayer sp) StageTask.checkStages(sp);
@@ -29,7 +32,7 @@ public class KubeJSIntegration extends KubeJSPlugin {
 			if (event.getPlayer() instanceof ServerPlayer sp) StageTask.checkStages(sp);
 		});
 
-		FTBXModCompat.LOGGER.info("FTB Quests: Enabled KubeJS integration");
+		FTBXModCompat.LOGGER.info("[FTB Quests] Enabled KubeJS integration");
 	}
 
 	@Override
@@ -45,6 +48,10 @@ public class KubeJSIntegration extends KubeJSPlugin {
 	@Override
 	public void registerEvents() {
 		FTBQuestsKubeJSEvents.EVENT_GROUP.register();
+	}
+
+	private static void onCustomFilterItem(CustomFilterDisplayItemsEvent event) {
+		FTBQuestsKubeJSEvents.CUSTOM_FILTER_ITEM.post(ScriptType.CLIENT, new CustomFilterItemEventJS(event));
 	}
 
 	public static EventResult onCustomTask(CustomTaskEvent event) {